% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean.R
\name{pmean}
\alias{pmean}
\title{Centering on person-means}
\usage{
pmean(ID, var)
}
\arguments{
\item{ID}{name of ID variable}

\item{var}{name of variable to be centered}
}
\value{
A column in your dataframe (with person-level means)
}
\description{
This function allows you calculate person-level means. This will create a level-2 variable that can be used in tandem with person-centered means. This is useful if you are interested in both the within-person and between-person effects.
}
\examples{
df <- data.frame(ID = c(1, 1, 2, 2), var = c(1, 3, 5, 9))
df$personMEAN <- pmean(df$ID, df$var)
}
\keyword{centering}
