% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcenter.R
\name{gcenter}
\alias{gcenter}
\title{Centering on grand-means}
\usage{
gcenter(var)
}
\arguments{
\item{var}{name of variable to be centered}
}
\value{
A column in your dataframe (with grand-mean centered data)
}
\description{
This function allows you to center on grand-means.
}
\examples{
\dontrun{data$centeredVAR<-gcenter(data$var)}
}
\keyword{centering}
