% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcenter.R
\name{pcenter}
\alias{pcenter}
\title{Centering on person-means}
\usage{
pcenter(ID, var)
}
\arguments{
\item{ID}{name of ID variable}

\item{var}{name of variable to be centered}
}
\value{
A column in your dataframe (with person-centered data)
}
\description{
This function allows you to center on person-means (also called "centering within clusters")
}
\examples{
\dontrun{data$centeredVAR<-pcenter(data$ID,data$var)}
}
\keyword{centering}
